+++ loaders/target-gcs/target_gcs/helpers/__init__.py
@@ -1,9 +1,13 @@
 from .json_parsing import _json_default
 from .partition_path import get_partition_path_from_record
-from .partition_schema import validate_partition_date_field_schema
+from .partition_schema import (
+    validate_partition_date_field_schema,
+    validate_partition_fields_schema,
+)
 
 __all__ = [
     "get_partition_path_from_record",
     "_json_default",
     "validate_partition_date_field_schema",
+    "validate_partition_fields_schema",
 ]